[refactor] 공통 기능 정리
