+++ src/main/java/potatoes/server/chat/stomp/StompUserPrincipal.java
@@ -1,4 +1,4 @@
-package potatoes.server.utils.stomp;
+package potatoes.server.chat.stomp;
 
 import java.security.Principal;
