+++ src/main/java/potatoes/server/infra/s3/S3Config.java
@@ -1,4 +1,4 @@
-package potatoes.server.utils.s3;
+package potatoes.server.infra.s3;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
