+++ src/main/java/potatoes/server/utils/constant/LocationType.java
@@ -1,4 +1,4 @@
-package potatoes.server.constant;
+package potatoes.server.utils.constant;
 
 public enum LocationType {
 	건대입구, 을지로3가, 신림, 홍대입구
