+++ src/main/java/potatoes/server/utils/constant/TravelStatus.java
@@ -1,4 +1,4 @@
-package potatoes.server.constant;
+package potatoes.server.utils.constant;
 
 public enum TravelStatus {
 	UPCOMING,
